+++ Assets/Scripts/Treasure.cs
@@ -4,15 +4,15 @@ using System.Collections;
 public class Treasure
 {
     public int id;
-    public float lat;
-    public float lng;
+    public double lat;
+    public double lng;
     public int type;
     
     public Treasure(int id, double lat, double lng, int type)
     {
         this.id = id;
-        this.lat = (float)lat;
-        this.lng = (float)lng;
+        this.lat = lat;
+        this.lng = lng;
         this.type = type;
     }
 }
